+++ zerver/views/user_groups.py
@@ -61,7 +61,7 @@ def add_user_group(
     can_manage_group: Json[int | AnonymousSettingGroupDict] | None = None,
     can_mention_group: Json[int | AnonymousSettingGroupDict] | None = None,
 ) -> HttpResponse:
-    user_profiles = user_ids_to_users(members, user_profile.realm)
+    user_profiles = user_ids_to_users(members, user_profile.realm, allow_deactivated=False)
     name = check_user_group_name(name)
 
     group_settings_map = {}
@@ -254,9 +254,8 @@ def notify_for_user_group_subscription_changes(
         if recipient_user.is_bot:
             # Don't send notification message to bots.
             continue
-        if not recipient_user.is_active:
-            # Don't send notification message to deactivated users.
-            continue
+
+        assert recipient_user.is_active
 
         with override_language(recipient_user.default_language):
             if send_subscription_message:
@@ -306,7 +305,7 @@ def add_members_to_group_backend(
             user_group_id, user_profile, permission_setting="can_manage_group"
         )
 
-    member_users = user_ids_to_users(members, user_profile.realm)
+    member_users = user_ids_to_users(members, user_profile.realm, allow_deactivated=False)
     existing_member_ids = set(
         get_direct_memberships_of_users(user_group.usergroup_ptr, member_users)
     )
@@ -337,7 +336,7 @@ def remove_members_from_group_backend(
     user_group_id: int,
     members: list[int],
 ) -> HttpResponse:
-    user_profiles = user_ids_to_users(members, user_profile.realm)
+    user_profiles = user_ids_to_users(members, user_profile.realm, allow_deactivated=False)
     user_group = access_user_group_for_update(
         user_group_id, user_profile, permission_setting="can_manage_group"
     )